Introduce useCache hook
