+++ frontend/src/utils/isDateExpired.ts
@@ -1,0 +1,9 @@
+/**
+ *
+ * @param date The date to be compared
+ * @param expiresIn data valid for N milliseconds
+ * @returns boolean
+ */
+export function isDateExpired(date: Date, expiresIn: number) {
+  return !(Math.abs(Number(new Date()) - Number(date)) < expiresIn);
+}
